feat: prepare offline desktop development

+++ app/model/purchase_line.model.go
@@ -3,25 +3,25 @@ package model
 import "time"
 
 type PurchaseLine struct {
-	ID           int64     `json:"id"`
-	EventID      int64     `json:"event_id"`
-	ItemID       int64     `json:"item_id"`
-	Quantity     float64   `json:"quantity"`
-	UnitCost     int64     `json:"unit_cost"`
-	CreatedAt    time.Time `json:"created_at"`
+	ID        int64     `json:"id"`
+	EventID   int64     `json:"event_id"`
+	ItemID    int64     `json:"item_id"`
+	Quantity  float64   `json:"quantity"`
+	UnitCost  int64     `json:"unit_cost"`
+	CreatedAt time.Time `json:"created_at"`
 }
 
 type PurchaseLineInsertDTO struct {
-	EventID      int64     `json:"event_id"`
-	ItemID       int64     `json:"item_id"`
-	Quantity     float64   `json:"quantity"`
-	UnitCost     int64     `json:"unit_cost"`
+	EventID  int64   `json:"event_id"`
+	ItemID   int64   `json:"item_id"`
+	Quantity float64 `json:"quantity"`
+	UnitCost int64   `json:"unit_cost"`
 }
 
 type CreatePurchaseWithLinesInput struct {
-	ItemID       int64     `json:"item_id"`
-	Quantity     float64   `json:"quantity"`
-	UnitCost     float64   `json:"unit_cost"`
+	ItemID   int64   `json:"item_id"`
+	Quantity float64 `json:"quantity"`
+	UnitCost int64   `json:"unit_cost"`
 }
 
 // type PurchaseLineUpdateDTO struct {
